Split off nodelist header components to nodelist.h
